pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "MotiTheWizerd"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.0,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "google-adk (>=1.21.0,<2.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
